Clean: Relocate STEX AffineTransforms
