+++ src/dewarping/CMakeLists.txt
@@ -12,10 +12,6 @@ SET(
         DistortionModel.cpp DistortionModel.h
         DistortionModelBuilder.cpp DistortionModelBuilder.h
         RasterDewarper.cpp RasterDewarper.h
-        STEX_AbstractImageTransform.h
-        STEX_AffineImageTransform.cpp STEX_AffineImageTransform.h
-        STEX_AffineTransform.cpp STEX_AffineTransform.h
-        STEX_AffineTransformedImage.cpp STEX_AffineTransformedImage.h
         STEX_GaussBlur.cpp STEX_GaussBlur.h
         STEX_HomographicTransform.h
         STEX_PlugHoles.h
