+++ src/dewarping/DewarpingImageTransform.cpp
@@ -22,7 +22,7 @@
 #include "STEX_ToVec.h"
 #include "ToLineProjector.h"
 #include "math/LineBoundedByPolygon.h"
-#include "STEX_AffineTransformedImage.h"
+#include "imageproc/AffineTransformedImage.h"
 #include <Eigen/Core>
 #include <Eigen/LU>
 #include <boost/optional.hpp>
