+++ src/dewarping/TextLineSegmenter.cpp
@@ -28,8 +28,8 @@
 #include "foundation/Grid.h"
 #include "foundation/MinMaxAccumulator.h"
 #include "LineBoundedByRect.h"
-#include "STEX_AffineTransformedImage.h"
-#include "STEX_AffineTransform.h"
+#include "imageproc/AffineTransformedImage.h"
+#include "imageproc/AffineTransform.h"
 #include "imageproc/Grayscale.h"
 #include "imageproc/GrayImage.h"
 #include "STEX_GaussBlur.h"
