+++ src/dewarping/TextLineTracer.cpp
@@ -28,8 +28,8 @@
 #include "math/ToLineProjector.h"
 #include "math/SidesOfLine.h"
 #include "math/LineBoundedByRect.h"
-#include "STEX_AffineTransformedImage.h"
-#include "STEX_AffineTransform.h"
+#include "imageproc/AffineTransformedImage.h"
+#include "imageproc/AffineTransform.h"
 #include "STEX_GaussBlur.h"
 #include "imageproc/Binarize.h"
 #include "imageproc/BinaryImage.h"
