+++ src/imageproc/AbstractImageTransform.h
